# docs/swagger.yaml
definitions:
  api.Message:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
  api.Meta:
    properties:
      filtered_count:
        type: integer
      total_count:
        type: integer
    type: object
  api.SwgErrResponse:
    properties:
      error:
        type: string
      messages:
        items:
          $ref: '#/definitions/api.Message'
        type: array
    type: object
...
  api.SwgResponse-array_userview_SimpleView:
    properties:
      data:
        items:
          $ref: '#/definitions/userview.SimpleView'
        type: array
      messages:
        items:
          $ref: '#/definitions/api.Message'
        type: array
      meta:
        $ref: '#/definitions/api.Meta'
    type: object
  api.SwgResponse-array_userview_View:
    properties:
      data:
        items:
          $ref: '#/definitions/userview.View'
        type: array
      messages:
        items:
          $ref: '#/definitions/api.Message'
        type: array
      meta:
        $ref: '#/definitions/api.Meta'
    type: object
...
  api.SwgResponse-reservationview_DetailedView:
    properties:
      data:
        $ref: '#/definitions/reservationview.DetailedView'
      messages:
        items:
          $ref: '#/definitions/api.Message'
        type: array
      meta:
        $ref: '#/definitions/api.Meta'
    type: object
  api.SwgResponse-userview_View:
    properties:
      data:
        $ref: '#/definitions/userview.View'
      messages:
        items:
          $ref: '#/definitions/api.Message'
        type: array
      meta:
        $ref: '#/definitions/api.Meta'
    type: object
...
  peopleparams.CreateParams:
    properties:
      full_name:
        type: string
      preferred_name:
        type: string
    type: object
...
  personview.View:
    properties:
      full_name:
        type: string
      id:
        type: integer
      language_preference:
        type: string
      preferred_name:
        type: string
    type: object
...
  reservationview.BookView:
    properties:
      author:
        type: string
      genre:
        type: string
      id:
        type: integer
      isbn:
        type: string
      language:
        type: string
      publication_date:
        type: string
      publisher:
        type: string
      title:
        type: string
    type: object
  reservationview.DetailedView:
    properties:
      book:
        $ref: '#/definitions/reservationview.BookView'
      book_id:
        type: integer
      id:
        type: integer
      reservation_date:
        type: string
      status:
        type: string
      user:
        $ref: '#/definitions/userview.View'
      user_id:
        type: integer
    type: object
  userparams.CreateParams:
    properties:
      email:
        type: string
      password:
        type: string
      person_attributes:
        $ref: '#/definitions/peopleparams.CreateParams'
      username:
        type: string
    type: object
...
  userview.SimpleView:
    properties:
      id:
        type: integer
      username:
        type: string
    type: object
  userview.View:
    properties:
      abilities:
        items:
          type: string
        type: array
      email:
        type: string
      id:
        type: integer
      person_attributes:
        $ref: '#/definitions/personview.View'
      username:
        type: string
    type: object
info:
  contact: {}
  title: Library Mangement System API
paths:
...
  /api/v1/reservation/{reservation_id}/:
    delete:
      consumes:
      - '*/*'
      description: Deletes a reservation for a book
      parameters:
      - description: Book ID for reservation
        in: path
        name: book_id
        required: true
        type: integer
      - description: reservation ID to delete
        in: path
        name: reservation_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SwgResponse-reservationview_DetailedView'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.SwgErrResponse'
      summary: Delete a reservation
      tags:
      - reservation
    get:
      consumes:
      - '*/*'
      description: Reads a reservation for a book
      parameters:
      - description: reservation ID to read
        in: path
        name: reservation_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SwgResponse-reservationview_DetailedView'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.SwgErrResponse'
      summary: Read a reservation
      tags:
      - reservation
...
  /api/v1/user:
    get:
      consumes:
      - '*/*'
      description: List users depending on the collection query
      parameters:
      - default: 0
        description: Offset for pagination
        in: query
        name: offset
        type: integer
      - default: 25
        description: Limit for pagination
        in: query
        name: limit
        type: integer
      - default: 1
        description: Filter by user ID
        in: query
        name: filter[user_id]
        type: integer
      - default: '"created_at"'
        description: Sort by column name
        in: query
        name: sortBy
        type: string
      - default: '"desc"'
        description: Order by asc or desc
        in: query
        name: orderBy
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SwgResponse-array_userview_View'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.SwgErrResponse'
      summary: List users
      tags:
      - user
  /api/v1/user/:
    post:
      consumes:
      - application/json
      description: Creates a new user in the system
      parameters:
      - description: User creation request
        in: body
        name: createuserparam
        required: true
        schema:
          $ref: '#/definitions/userparams.CreateParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SwgResponse-userview_View'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.SwgErrResponse'
      summary: Create a new user
      tags:
      - auth
...
  /api/v1/user/autocomplete:
    get:
      consumes:
      - '*/*'
      description: Autocomplete 10 users based on input value
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SwgResponse-array_userview_SimpleView'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.SwgErrResponse'
      summary: Autocomplete 10 users based on input value
      tags:
      - user
...
swagger: "2.0"